# modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/rest-openapi.yaml
components:
...
    schemas:
        Discount:
            properties:
                active:
                    example: true
                    type: boolean
                couponCode:
                    example: SAVE20
                    type: string
                customFields:
                    additionalProperties: true
                    type: object
                discountAccountGroups:
                    items:
                        $ref: "#/components/schemas/DiscountAccountGroup"
                    type: array
                discountCategories:
                    items:
                        $ref: "#/components/schemas/DiscountCategory"
                    type: array
                discountProducts:
                    items:
                        $ref: "#/components/schemas/DiscountProduct"
                    type: array
                discountRules:
                    items:
                        $ref: "#/components/schemas/DiscountRule"
                    type: array
                displayDate:
                    format: date
                    type: string
                expirationDate:
                    format: date
                    type: string
                externalReferenceCode:
                    example: AB-34098-789-N
                    type: string
                id:
                    example: 30130
                    format: int64
                    minimum: 0
                    type: integer
                limitationTimes:
                    example: 0
                    format: int32
                    minimum: 0
                    type: integer
                limitationType:
                    example: unlimited
                    type: string
                maximumDiscountAmount:
                    example: 25
                    format: bigdecimal
                    minimum: 0
                    type: number
                neverExpire:
                    example: true
                    type: boolean
                numberOfUse:
                    example: 5
                    format: int32
                    minimum: 0
                    type: integer
                percentageLevel1:
                    example: 20
                    format: bigdecimal
                    minimum: 0
                    type: number
                percentageLevel2:
                    example: 0
                    format: bigdecimal
                    minimum: 0
                    type: number
                percentageLevel3:
                    example: 0
                    format: bigdecimal
                    minimum: 0
                    type: number
                percentageLevel4:
                    example: 0
                    format: bigdecimal
                    minimum: 0
                    type: number
                target:
                    example: subtotal
                    type: string
                title:
                    example: 20% Off
                    type: string
                useCouponCode:
                    example: true
                    type: boolean
                usePercentage:
                    example: true
                    type: boolean
            required:
                - limitationType
                - target
                - title
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            scheme: bearer
            type: http
        OAuth2:
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/authorize
                    scopes:
                        HeadlessCommerceAdminPricing.admin: Grants access to admin operations
                        HeadlessCommerceAdminPricing.read: Grants read access
                        HeadlessCommerceAdminPricing.write: Grants write access
                    tokenUrl: /oauth/token
            type: oauth2
info:
    contact:
        email: team-commerce@liferay.com
        name: Commerce Team
    description:
        "Liferay Commerce Admin Pricing API. A Java client JAR is available for use with the group ID 'com.liferay',
        artifact ID 'com.liferay.headless.commerce.admin.pricing.client', and version '1.0.15'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: Liferay Commerce Admin Pricing API
    version: v1.0
openapi: 3.0.0
paths:
...
    /discounts:
        get:
            operationId: getDiscountsPage
            parameters:
                - in: query
                  explode: true
                  name: page
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  explode: true
                  name: pageSize
                  required: false
                  schema:
                      type: integer
                  style: form
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Discount"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Discount"
                                type: array
                    description:
                        Successful operation
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminPricing.read
            summary: Gets a list of Discount.
            tags:
                - Discount
        post:
            operationId: postDiscount
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Discount"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Discount"
                required: true
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Discount"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Discount"
                    description:
                        Created
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminPricing.write
            summary: Creates or Updates a Discount.
            tags:
                - Discount
    /discounts/by-externalReferenceCode/{externalReferenceCode}:
        delete:
            operationId: deleteDiscountByExternalReferenceCode
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            responses:
                204:
                    description:
                        No content
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminPricing.write
            summary: Deletes a Discount by external reference code.
            tags:
                - Discount
        get:
            operationId: getDiscountByExternalReferenceCode
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Discount"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Discount"
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminPricing.read
            summary: Gets a Discount by external reference code.
            tags:
                - Discount
        patch:
            operationId: patchDiscountByExternalReferenceCode
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Discount"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Discount"
                required: true
            responses:
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminPricing.write
            summary: Updates a Discount by external reference code.
            tags:
                - Discount
...
servers:
    - description:
          Liferay Commerce API server
      url: https://commerce.{environment}.liferay.com/commerce-admin-pricing/{basePath}
      variables:
          basePath:
              default: v1.0
          environment:
              default: api
              enum:
                  - api
                  - api.dev
                  - api.uat
    - description:
          Liferay Commerce Local Development (Hosts file)
      url: "{protocol}://{host}:{port}/o/commerce-admin-pricing/{basePath}"
      variables:
          basePath:
              default: v1.0
          host:
              default: test.dev
          port:
              default: "8080"
              enum:
                  - "443"
                  - "8080"
                  - "8443"
          protocol:
              default: http
              enum:
                  - http
                  - https
    - description:
          Liferay Commerce Local Development with Ngrok
...
      variables:
...